+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Date format: DD/MM/YYYY
 
-## [next]
+## [4.0.0-pre.1] - Materials and Pickers - [29/06/2022]
 
 - Exposed private properties that makes it easier to create custom panes for `NavigationView` ([#365](https://github.com/bdlukaa/fluent_ui/issues/365)):
   - `kCompactNavigationPaneWidth`